+++ anybox/recipe/openerp/base.py
@@ -174,7 +174,7 @@ class BaseRecipe(object):
         """
         self.version_wanted = self.options.get('version')
         if self.version_wanted is None:
-            raise ValueError('You must specify the version')
+            raise UserError('You must specify the version')
 
         self.preinstall_version_check()
 
@@ -185,9 +185,8 @@ class BaseRecipe(object):
             major_wanted = self.version_wanted[:3]
             pattern = self.archive_filenames[major_wanted]
             if pattern is None:
-                raise ValueError(
-                    'OpenERP version %r'
-                    'is not supported' % self.version_wanted)
+                raise UserError('OpenERP version %r'
+                                'is not supported' % self.version_wanted)
 
             self.archive_filename = pattern % self.version_wanted
             self.archive_path = join(self.downloads_dir, self.archive_filename)
@@ -210,7 +209,7 @@ class BaseRecipe(object):
             self.sources[main_software] = ('downloadable', (url, None))
         elif type_spec == 'nightly':
             if len(version_split) != 3:
-                raise ValueError(
+                raise UserError(
                     "Unrecognized nightly version specification: "
                     "%r (expecting series, number) % version_split[1:]")
             self.nightly_series, self.version_wanted = version_split[1:]
@@ -488,7 +487,7 @@ class BaseRecipe(object):
 
             split = line.split()
             if len(split) > 2:
-                raise ValueError("Invalid revisions line: %r" % line)
+                raise UserError("Invalid revisions line: %r" % line)
 
             # addon or main software
             if len(split) == 2:
@@ -505,8 +504,8 @@ class BaseRecipe(object):
                 continue
 
             if source[0] in ('downloadable', 'local'):
-                raise ValueError("In revision line %r : can't fix a revision "
-                                 "for non-vcs source" % line)
+                raise UserError("In revision line %r : can't fix a revision "
+                                "for non-vcs source" % line)
 
             logger.info("%s will be on revision %r", local_path, revision)
             self.sources[local_path] = ((source[0], (source[1][0], revision))
@@ -548,7 +547,7 @@ class BaseRecipe(object):
             manifest_pre_v6 = os.path.join(addons_dir, '__terp__.py')
             if os.path.isfile(manifest) or os.path.isfile(manifest_pre_v6):
                 if loc_type == 'local':
-                    raise ValueError(
+                    raise UserError(
                         "Local addons line %r should refer to a directory "
                         "containing addons, not to a standalone addon. "
                         "The recipe can perform automatic creation of "
@@ -699,7 +698,7 @@ class BaseRecipe(object):
 
         if ((freeze_to is not None or extract_downloads_to is not None)
                 and not self.offline):
-            raise ValueError("To freeze a part, you must run offline "
+            raise UserError("To freeze a part, you must run offline "
                              "so that there's no modification from what "
                              "you just tested. Please rerun with -o.")