+++ anybox/recipe/openerp/server.py
@@ -6,6 +6,7 @@ import shutil
 import logging
 import subprocess
 import zc.buildout
+from zc.buildout import UserError
 from anybox.recipe.openerp import devtools
 from anybox.recipe.openerp.base import BaseRecipe
 
@@ -237,7 +238,7 @@ conf = openerp.tools.config
 
             naming = line[0].split('=')
             if not naming or len(naming) > 2:
-                raise ValueError("Invalid script specification %r" % line[0])
+                raise UserError("Invalid script specification %r" % line[0])
             elif len(naming) == 1:
                 name = '_'.join((naming[0], self.name))
             else:
@@ -254,7 +255,7 @@ conf = openerp.tools.config
                 elif token.startswith(arg_prefix):
                     desc['arguments'] = token[len(arg_prefix):]
                 else:
-                    raise ValueError(
+                    raise UserError(
                         "Invalid token for script %r: %r" % (name, token))
 
     def _get_or_create_script(self, entry, name=None):